rename: performanceDetailHandlers 파일명 변경
