+++ src/mocks/browser.ts
@@ -1,5 +1,5 @@
 import { setupWorker } from 'msw/browser';
-import { performanceDetailHandlers } from './handlers/performanceDetail';
+import { performanceDetailHandlers } from './handlers/performanceDetailHandlers';
 import { profileHandlers } from './handlers/profileHandlers';
 
 export const worker = setupWorker(
